Allow enabling or disabling resources
This change adds in the concept of variables file, which we can split
out and load depending on the `DEPLOY_ENV` envvar. My reason for putting
into this file is that we can disable certain resources by default, and
then load additional resources per environment if we need to, rather
than having to specify a long list of envvars in the Makefile process.

It uses the concept of the `count` function which essentially creates
the number of resources specified, but is being used as a replacement
for a conditional (ref
https://github.com/hashicorp/terraform/issues/1604).

I have set the variable out that by default it is disabled, and then if
we wish to enable the resource we specifically add the variable value in
a per environment variables file, and this could be done vice versa
depending or whatever our preference ends up being.

This creates the "uktt_data" bucket in Production only.