Only allow users to access AWS from the office.
This is shamelessly stolen from the excellent work done
by @alextomlins and the PaaS team.

https://github.gds/government-paas/aws-account-wide-terraform/commit/b5e627b19b84303eac5e8c2f4f0516a1de6cc6b4

"This adds a policy to all users to restrict access to office IPs. This
is to offer some defence in depth in case an access key is leaked.

The effect of this policy is that all access to AWS using access keys
will have to come from a trusted IP. All access from VMs within AWS will
therefore have to use instance profiles."