Allow the Rakefile to do project based runs.
We now treat things as either 'legacy', which uses the old state bucketname
and can be run via the apply_legacy task or project based which uses the
project name to isolate things like the state file. The project layout will
be fully documented in an ADR once testing is complete.