Add the ability to run spec files in our given structure
This will only run the spec files for the resources that should have run.
If a project has a resources and a resources/integration it will run the specs
from both spec/ and spec/integration. If you run the same tests but with the environment
set to stage the integration specs will no longer run but the main ones will.