locks down ami_builder IAM user to a single VPC
We want to ensure that the IAM user can create/destroy instances
in a specific VDC only to protect other VDCs from accidental or
malicious shutdown of instances.  IAM permissions are quite complicated
and unintuitive. After much trial and error, we successfully followed
the instructions in this Amazon post: http://amzn.to/1ouXPo0

In essence:
1) create a role with no policy attached (VPCLockDown)
2) In the user policy, give the user access to:
  - assign the role to instances
  - control instances with the role
  - bring up instances in a specific VPC (only)
  - global permissions to describe things (to discover resources)

Whereas it's not really obvious why we need to create an instance role,
that is the key to making the user policy work.