Add email_alert_notifications project
This project sets up an S3 bucket and a lambda function that parses
files and renames them with a prefix (currently only
`travel-advice-alerts`) and the `govuk_request_id` parsed from the email
body.

This allows us to more reliably verify that email alerts have
been sent when a travel advice update has been published.

An SES domain and ruleset are also required but can't currently be configured
via Terraform.