Pull object versionid
There was a requirement to be able to pull in the latest VersionID of a
specific object in a bucket, with a prerequisite that it was triggered
by a Jenkins deployment job. This changes seeks to add that
functionality by allowing the addition of an environmental variable in
the deploy job named "TF_VAR_LAMBDA_FILENAME".

I updated this in the project it's used so it could be used more
generically in the future. There might be a bit of extra work if we wish
to override the VersionID for whatever reason.