Update location of templates
Terraform only works with stuff when it is quoted from the relative
directory. Our rake task pushes everything into a temporary directory
and then feeds that in during the plan/apply system call. In the
Terraform code we specified a relative path, so to fix this I have
created a new templates directory with the project name underneath it.

Another solution would be to pass in the "tmpdir" variable in the Rake
task as env var, but I thought this was not as tidy as just having a
separate directory especially for templates.