Change Lambda app bucket to be public
We're having problems deploying the email_alert_notifcations project
because we think that it's using a different role to access this bucket
rather the credentials we're passing in. Making it public would allow us
to access the bucket, and seeing as the code is public anyway it's not a
big deal.