Allow Lambda to make requests with IP restrictions
We have code in an S3 bucket that we want Lambda to be able to download
and run. This doesn't work at the moment because of our IP address
restrictions.

This commit adds an exception to the IP address restrictions so that
any Lambda function in eu-west-1 from our account can make requests
from any IP address.