Update permissions of RW user
We have some functions which moves stuff around in buckets using the
private_s3_bucket module. It seems that without some permissions the
user is unable to do stuff like move objects around and (more obviously)
delete objects for the move. I tested manually and it seems the ACL
permissions are required.