Extract `terraform get` to outside the loop.
There is no reason to run the `get` on each file. It can also cause errors such as

    config: unknown resource 'aws_iam_user.foo' referenced in variable bar

when cross file resources are accessed in an incorrect order. By running the get
when all resources are in the same directory we should reduce the false postives and only
break when resources are acually missing.