Adjust govuk-backup S3 permissions
The initial backup worked, but in trying to run future backups duplicity
threw a 403 when trying to connect to the S3 bucket. I believe it may
have been trying to move or do something with the files in the bucket.

It's easiest to just allow full permission by that user. It only has
access to that bucket.