Remove the unused account_id variable
The `account_id` variable defined in `config/common_variables` is not
used in any current projects. The only usage is in
`old-projects/packer_ami_builder` which is deprecated.

Where the `account_id` of the current user is needed the
data resource `aws_caller_identity`'s attribute `current.account_id`
should be preferred.