Remove lifecycle and versioning
Duplicity has lots of strange errors related to orphaned files and seems
to run a full backup each time it runs, and doesn't leave any old
backups.

I think this may be related to the lifecycle rule of a single day.
Duplicity itself should deal with the clean up, and I feel that having
both versioning and lifecycles enabled on the bucket complicate matters
unneccessarily.