Allow use of multiple providers
This separates out common configuration variables from provider
specific ones. This allows us to use, for example, Google Compute
Engine (GCE) as well as AWS.

To achieve this provider specific configurations are separated into
their own directories. By default the provider is assumed to be AWS
but any project prefixed with `gce_` is treated as GCE.