Add network and security stack for GOVUK Training environment
Terraform resources to create a basic network stack and policies
for our network environment. The training VM consumes the remote
state file of this stack to locate its subnet, instance profile
and security groups.

We should update this stack later to move everything to a private
subnet and use ELBs, EIPs, etc.