Add S3 mirror policy for Pingdom IPs
We want to add Pingdom checks for our S3 mirrors so we need to whitelist
the probe IPs. These IPs are likely to change, but they can be pulled
automatically as explained in
https://help.pingdom.com/hc/en-us/articles/203682601-Pingdom-probe-servers-IP-addresses

To manage these IPs automatically, there is a bash script that populates
a Terraform external data source. It's been kept very simple so we don't
have to install dependencies just for this on the servers. There is a limitation
in Terraform that it can only pull string values from external data sources
https://github.com/hashicorp/terraform/issues/12256
so we need to manipulate the string in the Terraform code.