Add "public read" policy to Asset Manager S3 bucket
We want to experiment with having Asset Manager redirect asset URLs to
objects on S3 - see this pull request [1] for details. Note that this
experimental Asset Manager behaviour is disabled by default, but can be
enabled using environment variables or on a per-request basis using a
query string parameter [2].

Previously the default bucket policy did not allow public read access,
but now that we want to redirect requests and have S3 serve the asset to
the end-user, we need to add an explicit policy to the bucket to allow
public read access.

We considered setting the policy on individual objects, but decided it
would be simpler to apply it to the bucket as a whole and then create a
separate bucket if we need some objects to have a more restrictive
policy.

Note that we will want to create a DNS CNAME and rename the S3 bucket at
some point to get the full effect, but this feels like a useful first
step.

[1]: alphagov/asset-manager#112
[2]: https://github.com/alphagov/asset-manager/blob/master/README.md#assets-on-s3