Make Terraform state S3 bucket configurable
These changes allow you to run the standard Rake tasks to plan/apply
changes even if you don't have access to the GOV.UK S3 buckets which
store Terraform state for the various environments.

This isn't necessarily the best way to solve this problem, because the
Terraform state S3 bucket name is hard-coded elsewhere in the repo.
However, given that we understand that this repo is likely to disappear
in the near future, it feels like a pragmatic solution to the problem
we were having.