Only apply public-read bucket policy in integration
I've tested this against my own AWS account and can confirm that the
bucket policy is only applied in the integration environment (i.e. when
`DEPLOY_ENV=integration`). This hopefully addresses the infrastructure
team's concern about making this change in staging/production without us
also configuring logging and a CDN. While we plan to do both of those
things, we don't want them to prevent us making progress with testing
the effect of this change in integration.