Remove public-read policy from asset-manager S3 bucket
We added this bucket policy in govuk-terraform-provisioning PR 130[1] so
that we could test the effect of redirecting to assets served on S3.

We've decided to pursue an alternative approach in the short term
(proxying requests via nginx) so we don't need this public policy set on
the bucket.

I've tested this using my local AWS account and confirmed that it
removed the policy as expected.

[1]: https://github.com/alphagov/govuk-terraform-provisioning/pull/130