Add Asset Manager backup S3 bucket in production
I want to setup cross-region replication for the Asset Manager S3
bucket. This new S3 bucket is intended to be used as the target for the
replication and this is why it has its region set to `eu-west-2` i.e. so
that it's different from the default region which is `eu-west-1` as
defined in `configs/common_variables.tf`.

Note also that I had to introduce a new AWS provider in order to use the
appropriate AWS API endpoint to create the backup S3 bucket.