Add variable for lifecycle
Update the incredibly useful "private_s3_bucket" module that we used
extensively until something happened and we moved away from using it.
When we want to create a bucket and user without faff it was infinitely
useful.

I believe we stopped using it due to lifecycle policies; these are
disabled by default, so unless specified, things in buckets should not
be deleted.

This commit adds a variable so we can specify the amount of days that
things should be deleted. Default is 30.