Allow syncing of assets between environments
The idea is that a script in the `env-sync-and-backup` will SSH onto a
production machine and use the `aws s3 sync` command line tool to
perform a nightly sync of the production assets S3 bucket to the staging
& integration assets S3 buckets.

* Adds a new IAM user to the Asset Manager project in the production
environment with minimal read access to the production assets S3 bucket
and minimal write access to the staging & integration assets S3 buckets.

* Adds a bucket policy to the staging & integration assets S3 buckets to
allow minimal write access to the new production IAM user mentioned
above.

* Note that, because each Terraform environment has its own separate AWS
account, we need to pass in the AWS account ID as a Terraform variable,
`production_aws_account_id`, in order to be able to construct the bucket
policy mentioned above.