Add s3:PutObjectAcl permission for production env sync user
The env-sync-and-backup project does a nightly sync of assets from the
production S3 bucket to the staging and integration buckets. This is
achieved using a env sync production user,
`govuk-assets-production-env-sync-user` which has permission to get, put
& delete objects in the target buckets. Unfortunately this meant the
sync-ed objects had their owner set to the env sync production user.
This in turn meant that the Asset Manager AWS user in staging &
integration had no permission to do anything with the sync-ed objects -
not even read them!

The way we've decided to fix this is to change the `aws s3 sync` command
used in the env-sync-and-backup project to use the `acl` option set to
`bucket-owner-full-control`. However, in order to be able to use this
option, we need to give the env sync production user permission to write
object ACLs, i.e. by adding the `s3:PutObjectAcl` action to the relevant
policy.